Update: Convert layout to RelativeLayout
Android Studio initially had layout set as a ConstraintLayout. The
example provided had a RelativeLayout but it would not build and caused
errors within Android Studio.  Used the Android Studio tools to set the
layout back to the original ConstraintLayout that Android Studio had
set. Using the convert tool, was able to change the layout to a working
RelativeLayout that now builds.