Add margins to TextViews and adjust textSize
Margins were added to the TextViews so they were not right up against
the screen borders. They now fall in line with the Android style guide
of using 16dp from the edge.

The text size was also a bit big at 32sp. The new size has been reduced
to 24sp for a cleaner look.